refactor(FTQ): make ftq a separate package and split ftq into multiple files (#4683)
No new feature added.
