+++ src/main/scala/xiangshan/backend/decode/DecodeStage.scala
@@ -29,7 +29,7 @@ import xiangshan.backend.fu.FuType
 import xiangshan.backend.fu.wrapper.CSRToDecode
 import yunsuan.VpermType
 import xiangshan.ExceptionNO.{illegalInstr, virtualInstr}
-import xiangshan.frontend.FtqPtr
+import xiangshan.frontend.ftq.FtqPtr
 
 class DecodeStageIO(implicit p: Parameters) extends XSBundle {
   // params alias
