+++ src/main/scala/xiangshan/frontend/ftq/FtqPtr.scala
@@ -1,0 +1,35 @@
+// Copyright (c) 2024-2025 Beijing Institute of Open Source Chip (BOSC)
+// Copyright (c) 2020-2025 Institute of Computing Technology, Chinese Academy of Sciences
+// Copyright (c) 2020-2021 Peng Cheng Laboratory
+//
+// XiangShan is licensed under Mulan PSL v2.
+// You can use this software according to the terms and conditions of the Mulan PSL v2.
+// You may obtain a copy of Mulan PSL v2 at: https://license.coscl.org.cn/MulanPSL2
+//
+// THIS SOFTWARE IS PROVIDED ON AN "AS IS" BASIS, WITHOUT WARRANTIES OF ANY KIND,
+// EITHER EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO NON-INFRINGEMENT,
+// MERCHANTABILITY OR FIT FOR A PARTICULAR PURPOSE.
+//
+// See the Mulan PSL v2 for more details.
+
+package xiangshan.frontend.ftq
+
+import chisel3._
+import org.chipsalliance.cde.config.Parameters
+import utility.CircularQueuePtr
+import xiangshan.XSCoreParamsKey
+
+class FtqPtr(entries: Int) extends CircularQueuePtr[FtqPtr](entries) {
+  def this()(implicit p: Parameters) = this(p(XSCoreParamsKey).FtqSize)
+}
+
+object FtqPtr {
+  def apply(f: Bool, v: UInt)(implicit p: Parameters): FtqPtr = {
+    val ptr = Wire(new FtqPtr)
+    ptr.flag  := f
+    ptr.value := v
+    ptr
+  }
+
+  def inverse(ptr: FtqPtr)(implicit p: Parameters): FtqPtr = apply(!ptr.flag, ptr.value)
+}
